+++ crates/zng/src/widget.rs
@@ -59,7 +59,7 @@ pub use zng_wgt::{
     OnNodeOpArgs, WeakWidgetFn, Wgt, WidgetFn, EDITORS, IS_MOBILE_VAR,
 };
 
-pub use zng_wgt_image::border::{border_img, border_img_repeat, border_img_fill, BorderRepeats};
+pub use zng_wgt_image::border::{border_img, border_img_fill, border_img_repeat, BorderRepeats};
 
 pub use zng_wgt_fill::{
     background, background_color, background_conic, background_fn, background_gradient, background_radial, foreground, foreground_color,