+++ examples/window/src/main.rs
@@ -27,11 +27,6 @@ fn app_main() {
 async fn main_window() -> window::WindowRoot {
     // WINDOWS.exit_on_last_close().set(false);
 
-    zng::image::IMAGES.limits().modify(|l| {
-        let l = l.to_mut();
-        l.allow_path = zng::image::PathFilter::allow_dir(zng::env::res(""));
-    });
-
     let window_vars = WINDOW.vars();
     let title = merge_var!(
         window_vars.actual_position(),